requirements.txt
alabaster==0.7.12
autopep8==1.6.0
babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
#dgl-cu113==0.8.2.post1
dglgo==0.0.1
docker-pycreds==0.4.0
docutils==0.18.1
gitdb==4.0.9
gitpython==3.1.27
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
isort==5.10.1
jinja2==3.1.2
joblib==1.1.0
littleutils==0.2.2
markupsafe==2.1.1
networkx==2.8.4
numpy==1.23.1
numpydoc==1.4.0
ogb==1.3.3
outdated==0.2.1
packaging==21.3
pandas==1.4.3
pathtools==0.1.2
pillow==9.2.0
pip==22.1.2
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pycodestyle==2.8.0
pydantic==1.9.1
pygcl==0.1.2
pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyyaml==6.0
requests==2.28.1
ruamel.yaml.clib==0.2.6
ruamel.yaml==0.17.21
scikit-learn==1.1.1
scipy==1.8.1
sentry-sdk==1.7.1
setproctitle==1.2.3
setuptools==61.2.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.1.0
toml==0.10.2

# torch==1.11.0+cu102
# torchaudio==0.11.0+cu102
# torchvision==0.12.0+cu102
# torch-cluster==1.6.0
# torch-geometric==2.0.4
# torch-scatter==2.0.9
# torch-sparse==0.6.14
# torch-spline-conv==1.2.1

tqdm==4.64.0
typer==0.6.1
typing-extensions==4.3.0
urllib3==1.26.10
wandb==0.12.21
wheel==0.37.1
zipp==3.8.1
